# [WIP] Immediate writes to underlying stream until buffer is filled