--- a/tests/BufferTest.php
+++ b/tests/BufferTest.php
@@ -37,7 +37,7 @@ class BufferTest extends TestCase
     public function testWrite()
     {
         $stream = fopen('php://temp', 'r+');
-        $loop = $this->createWriteableLoopMock();
+        $loop = $this->createLoopMock();
 
         $buffer = new Buffer($stream, $loop);
         $buffer->on('error', $this->expectCallableNever());
@@ -96,13 +96,11 @@ class BufferTest extends TestCase
 
     /**
      * @covers React\Stream\Buffer::write
-     * @covers React\Stream\Buffer::handleWrite
      */
     public function testWriteReturnsFalseWhenBufferIsFull()
     {
         $stream = fopen('php://temp', 'r+');
-        $loop = $this->createWriteableLoopMock();
-        $loop->preventWrites = true;
+        $loop = $this->createLoopMock();
 
         $buffer = new Buffer($stream, $loop);
         $buffer->softLimit = 4;
@@ -110,7 +108,6 @@ class BufferTest extends TestCase
         $buffer->on('error', $this->expectCallableNever());
 
         $this->assertTrue($buffer->write("foo"));
-        $loop->preventWrites = false;
         $this->assertFalse($buffer->write("bar\n"));
     }
 
@@ -123,6 +120,7 @@ class BufferTest extends TestCase
         $loop = $this->createLoopMock();
 
         $buffer = new Buffer($stream, $loop);
+        $buffer->listening = true;
         $buffer->softLimit = 3;
 
         $this->assertFalse($buffer->write("foo"));
@@ -147,7 +145,6 @@ class BufferTest extends TestCase
         //$buffer->on('close', $this->expectCallableOnce());
 
         $buffer->write('hello');
-        $buffer->handleWrite();
     }
 
     /**
@@ -172,21 +169,34 @@ class BufferTest extends TestCase
      * @covers React\Stream\Buffer::write
      * @covers React\Stream\Buffer::handleWrite
      */
-    public function testDrain()
+    public function testEmitsDrainIfBufferIsFullAndThenEmpties()
     {
         $stream = fopen('php://temp', 'r+');
-        $loop = $this->createWriteableLoopMock();
-        $loop->preventWrites = true;
+        $loop = $this->createLoopMock();
 
         $buffer = new Buffer($stream, $loop);
         $buffer->softLimit = 4;
         $buffer->on('error', $this->expectCallableNever());
         $buffer->on('drain', $this->expectCallableOnce());
 
+        $buffer->write("foobar");
+    }
+
+    /**
+     * @covers React\Stream\Buffer::write
+     * @covers React\Stream\Buffer::handleWrite
+     */
+    public function testDoesNotEmitDrainIfBufferIsNotFull()
+    {
+        $stream = fopen('php://temp', 'r+');
+        $loop = $this->createLoopMock();
+
+        $buffer = new Buffer($stream, $loop);
+        $buffer->softLimit = 4;
+        $buffer->on('error', $this->expectCallableNever());
+        $buffer->on('drain', $this->expectCallableNever());
+
         $buffer->write("foo");
-        $loop->preventWrites = false;
-        $buffer->listening = false;
-        $buffer->write("bar\n");
     }
 
     /**
@@ -196,8 +206,7 @@ class BufferTest extends TestCase
     public function testWriteInDrain()
     {
         $stream = fopen('php://temp', 'r+');
-        $loop = $this->createWriteableLoopMock();
-        $loop->preventWrites = true;
+        $loop = $this->createLoopMock();
 
         $buffer = new Buffer($stream, $loop);
         $buffer->softLimit = 2;
@@ -210,7 +219,6 @@ class BufferTest extends TestCase
 
         $buffer->listening = true;
         $this->assertFalse($buffer->write("foo"));
-        $loop->preventWrites = false;
         $buffer->listening = false;
         $buffer->write("\n");
 
@@ -282,7 +290,7 @@ class BufferTest extends TestCase
     public function testEndWithData()
     {
         $stream = fopen('php://temp', 'r+');
-        $loop = $this->createWriteableLoopMock();
+        $loop = $this->createLoopMock();
 
         $buffer = new Buffer($stream, $loop);
         $buffer->on('error', $this->expectCallableNever());
@@ -319,7 +327,7 @@ class BufferTest extends TestCase
     public function testWritingToClosedBufferShouldNotWriteToStream()
     {
         $stream = fopen('php://temp', 'r+');
-        $loop = $this->createWriteableLoopMock();
+        $loop = $this->createLoopMock();
 
         $buffer = new Buffer($stream, $loop);
         $buffer->close();
@@ -336,7 +344,7 @@ class BufferTest extends TestCase
     public function testErrorWhenStreamResourceIsInvalid()
     {
         $stream = fopen('php://temp', 'r+');
-        $loop = $this->createWriteableLoopMock();
+        $loop = $this->createLoopMock();
 
         $error = null;
 
@@ -364,7 +372,7 @@ class BufferTest extends TestCase
         }
 
         list($a, $b) = stream_socket_pair(STREAM_PF_UNIX, STREAM_SOCK_STREAM, STREAM_IPPROTO_IP);
-        $loop = $this->createWriteableLoopMock();
+        $loop = $this->createLoopMock();
 
         $error = null;
 
@@ -380,22 +388,6 @@ class BufferTest extends TestCase
 
         $this->assertInstanceOf('Exception', $error);
         $this->assertSame('Unable to write to stream: fwrite(): send of 3 bytes failed with errno=32 Broken pipe', $error->getMessage());
-    }
-
-    private function createWriteableLoopMock()
-    {
-        $loop = $this->createLoopMock();
-        $loop->preventWrites = false;
-        $loop
-            ->expects($this->any())
-            ->method('addWriteStream')
-            ->will($this->returnCallback(function ($stream, $listener) use ($loop) {
-                if (!$loop->preventWrites) {
-                    call_user_func($listener, $stream);
-                }
-            }));
-
-        return $loop;
     }
 
     private function createLoopMock()