--- a/tests/StreamTest.php
+++ b/tests/StreamTest.php
@@ -123,7 +123,7 @@ class StreamTest extends TestCase
     public function testWrite()
     {
         $stream = fopen('php://temp', 'r+');
-        $loop = $this->createWriteableLoopMock();
+        $loop = $this->createLoopMock();
 
         $conn = new Stream($stream, $loop);
         $conn->write("foo\n");
@@ -234,19 +234,6 @@ class StreamTest extends TestCase
         rewind($stream);
 
         $conn->handleData($stream);
-    }
-
-    private function createWriteableLoopMock()
-    {
-        $loop = $this->createLoopMock();
-        $loop
-            ->expects($this->once())
-            ->method('addWriteStream')
-            ->will($this->returnCallback(function ($stream, $listener) {
-                call_user_func($listener, $stream);
-            }));
-
-        return $loop;
     }
 
     private function createLoopMock()