@clue 2016-08-21
Afaict this has also been suggested in #15 by @mbonneau around 2 years ago.