@WyriHaximus 2016-08-21
This combined with #53 puts my benchmark at 3000MB/s where #53 already raised it to 2500MB/s from 260MB/s on my laptop (PHP 7).