@davidwdan 2016-10-19
Can we get this merged?  It bit me today and wasted many hours.