@WyriHaximus 2016-10-19
Agreed @davidwdan lets get this in. @clue what is holding us back from merging this and related issues?