@clue 2016-10-19
> Can we get this merged?

Absolutely! :+1: 

> It bit me today and wasted many hours.

Huh, care to elaborate?