@clue 2016-11-13
Rebased now that #55 is in.

I've moved this PR to the next milestone because it requires an update to not emit a `drain` event during an immediate write. This actually requires quite a few updates in our test suite because most test cases do not expect immediate writes and thus expect a `drain` event.