@clue 2016-12-19
Removing the milestone for now until the `drain` behavior is addressed.